// src/types/ActionTypes/ActionTypes.tsx
export enum ActionTypes {
  MOVE_TASK = "MOVE_TASK",
  ADD_TASK = "ADD_TASK",
  PUT_TASK = "PUT_TASK",
  ADD_COMMENT = "ADD_COMMENT",
  SUB_ADD_COMMENT = "SUB_ADD_COMMENT",
  ADD_SUBTASK = "ADD_SUBTASK",
  ADD_FILES = "ADD_FILES",
  FETCH_PROJECTS = "FETCH_PROJECTS",
  SET_FAIL = "SET_FAIL",
  SET_PROJECTS = "SET_PROJECTS",
  SET_LOAD = "SET_LOAD",
  UPDATE_PROJECT = "UPDATE_PROJECT",
  INIT_PROJECT_BACK = "INIT_PROJECT_BACK",
}